Created dominion chat and minor changes to parser and dominion.
Dominion chat is a new tool developed to receive messages through DNS.
It can be used for fun and for penetration testing purposes. It is under
heavy development because for now it only supports A questions.

Dominion 0.1.0 -> 0.2.0: Added client information to the ServerService trait.
Parser 0.2.0 -> 0.3.0: Implemented TryFrom<&'a str> for Name.